led matrix - resolved SPI chaining issues
Because store and show ping-pong shift-registers were used, the data coming out
of the MISO for the last LED in the chain was incorrect.  This is not strictly
an issue, because its passed the end of the chain.  However, this is now
changed to a perhaps better approach using only one shift-register.